superResolution/realesrgan/rknn/python/requirements.txt
# Real-ESRGAN Requirements for RKNN Conversion
# Python 3.10 required (conda environment: rknn310)

# Core dependencies
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.20.0
opencv-python>=4.5.0
Pillow>=8.0.0

# ONNX export and conversion
onnx>=1.10.0
onnxruntime>=1.10.0
onnx-simplifier>=0.4.0

# RKNN toolkit (should already be installed in rknn310 environment)
# rknn-toolkit2==2.3.0

# Image quality metrics (optional, for comparison)
# scikit-image>=0.18.0

# Note: Make sure you have activated the rknn310 conda environment:
# conda activate rknn310
# pip install -r requirements.txt
